+++ sw/simplex/simplex.h
@@ -3,13 +3,13 @@
 
 typedef struct simplex_t simplex_t;
 
-double simplex(int m, int n, double** a, double* b, double* c, double* x, double y);
+float simplex(int m, int n, float** a, float* b, float* c, float* x, float y);
 
-double xsimplex(int m, int n, double** a, double* b, double* c, double* x, double y, int* var, int h);
+float xsimplex(int m, int n, float** a, float* b, float* c, float* x, float y, int* var, int h);
 
-int initial(simplex_t* s, int m, int n, double** a, double* b, double* c, double* x, double y, int* var);
+int initial(simplex_t* s, int m, int n, float** a, float* b, float* c, float* x, float y, int* var);
 
-int init(simplex_t* s, int m, int n, double** a, double* b, double* c, double* x, double y, int* var);
+int init(simplex_t* s, int m, int n, float** a, float* b, float* c, float* x, float y, int* var);
 
 void prepare(simplex_t* s, int k);
